+++ frontend/src/components/pages/Auth.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import FormRegistration from "../FormRegistration";
+import FormRegistration from "../ui/FormRegistration";
 
 const Auth: React.FC = () => {
   return (
